feat: add named enums to graphql description, add deprecation directives
extends the existing UseEnumNames functionality to add the enum values
to the description of the graphql schema. This allows for clients of the
graphql schema to be aware of the values of the enums. Useful for
converting an enum to a nice display value without needing to manually
maintain a separate mapping from enum to display value.

adds a new `DeprecatedEnumValues` annotation. This allows for specifying
an array of enum values that should have the "deprecated" directive
added in the graphql schema.

in combination, these allow for a better client experience.